chore: finish the clippy::perf sweep (dead same-type casts) (#145)
* chore(core): remove dead same-type casts

Two clippy::unnecessary_cast findings in real (non-test) code:
AtomicU64::fetch_add already returns u64 (error_reporting.rs's event
counter); ClipInstance::source_out_secs/source_in_secs are already f64
(timeline/mod.rs's multicam angle-switch duration calc). Both casts
were dead code the compiler already elided -- zero runtime effect
either way, a correctness/clarity cleanup rather than a performance
fix (see perf/REPORT.md's own "Not fixed this pass" section for why
this category was originally scoped out of the perf pass; finishing
it here for consistency now that two instances were touched anyway).


* chore(ui): remove dead same-type casts

theme.rs's six CornerRadius::same(RADIUS_SM/RADIUS_MD as u8) calls and
draw.rs's transition-wedge width calc: RADIUS_SM/RADIUS_MD are already
u8, ClipInstance::transition_duration_secs is already f32. Same
dead-code cleanup as the companion core commit -- see that commit's
message for why this is filed separately from the perf pass itself.


* docs(perf): note the clippy::perf workspace sweep is now exhaustive

Cross-references the companion chore(core)/chore(ui) dead-cast commits
and records that every remaining clippy::perf-group warning in the
workspace is now either in test-only code (documented, deliberately
left) or a style-adjacent lint that doesn't actually concern runtime
cost (type_complexity, ptr_arg, too_many_arguments,
enum_variant_names) -- closing out this round of analysis rather than
leaving it open-ended.


---------

+++ crates/core/src/error_reporting.rs
@@ -891,7 +891,7 @@ fn next_id_hex() -> String {
         .map(|d| d.as_nanos() as u64)
         .unwrap_or(0);
     let pid = std::process::id() as u64;
-    let counter = EVENT_COUNTER.fetch_add(1, Ordering::Relaxed) as u64;
+    let counter = EVENT_COUNTER.fetch_add(1, Ordering::Relaxed);
     let mixed = splitmix64(nanos ^ (pid.wrapping_mul(0x9E3779B9)) ^ (counter << 32));
     format!("{mixed:016x}")
 }
